Reorganizing forms properly.
