+++ forms/__init__.py
@@ -1,0 +1,3 @@
+from flask_wtf import FlaskForm
+from wtforms import StringField, SubmitField, DecimalField, FieldList, FormField, DateField, SelectField, IntegerField
+from wtforms.validators import DataRequired
